requirements.txt
#PyPI
anndata==0.8.0
boto3==1.26.48
cell2location==0.1.3
scvi-tools==1.0.3

optax==0.1.7
chex==0.1.7
cellpose==2.1.1
distinctipy==1.2.2
frigidum==0.4.0
geopandas==0.12.2
hmmlearn==0.2.8
hydra-core==1.3.1
jax==0.4.14
jaxlib==0.4.14
light-size-constrained-clustering==0.0.7
matplotlib==3.6.2
numpy==1.22.4
pandas==1.5.2
patchify==0.2.3
pylint==2.15.10
pytest-cov==4.0.0
pytest-mock==3.10.0
pytest==7.2.0
python-dotenv==0.21.0
pytz==2022.7.1
PyYAML==6.0.1
rasterio==1.3.4
Rtree==1.0.1
scanpy==1.9.1
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.10.0
shapely==2.0.0
sphinx-rtd-theme==0.5.1
Sphinx==6.1.3
squidpy==1.2.3
tables==3.8.0
threadpoolctl==3.1.0
tifffile==2022.10.10
staintools==2.1.2
# spams==2.6.5.4
hdf5plugin==4.1.1
dash-bio==1.0.2
jinja2==3.1.2
holoviews==1.16.0
plotly==5.13.1
matplotlib==3.6.2
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
Pillow >= 10.0.1
metaflow==2.9.3
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
